Code Refactor
- Refactored index.js into multiple script files.
- Improved the 404 page.
- Improved Code Comments.
