--- routes/admin_pages.js
@@ -1,2 +1,0 @@
-const express = require('express');
-var router = express.Router();
