+++ views/404.hbs
@@ -1,0 +1,13 @@
+<head>
+    <link rel="stylesheet" href="/bootstrap.min.css" />
+    <link rel="stylesheet" href="/theme.css" />
+    <title>Page Not Found</title>
+</head>
+
+<body>
+    {{> head}}
+    <h1 class="txt-color text-center">Error 404</h1>
+    <h3 class="txt-color text-center">Page not found!</h3>
+
+    <p class="text-center mt-5"><a href="/" class="btn btn-info">Back to Homepage.</a></p>
+</body>
